initialize OneSignal only on client-side to prevent SSR errors and enhance notification handling
